migrate onConnection out of index
